restful/src/main/java/com/opensource/restful/shared/dto/EmailTypeDTO.java
package com.opensource.restful.shared.dto;

import java.io.Serializable;

@SuppressWarnings("serial")
public class EmailTypeDTO implements Serializable
{
    private long id;
    private String description;
    private boolean active;

    public long getId()
    {
        return id;
    }

    public void setId(long id)
    {
        this.id = id;
    }

    public String getDescription()
    {
        return description;
    }

    public void setDescription(String description)
    {
        this.description = description;
    }

    public boolean isActive()
    {
        return active;
    }

    public void setActive(boolean active)
    {
        this.active = active;
    }

    @Override
    public String toString()
    {
        return "EmailTypeDTO [id=" + id + ", description=" + description + ", active=" + active + "]";
    }

    @Override
    public int hashCode()
    {
        final int prime = 31;
        int result = 1;
        result = prime * result + (active ? 1231 : 1237);
        result = prime * result + ((description == null) ? 0 : description.hashCode());
        result = prime * result + (int) (id ^ (id >>> 32));
        return result;
    }

    @Override
    public boolean equals(Object obj)
    {
        if (this == obj)
            return true;
        if (obj == null)
            return false;
        if (getClass() != obj.getClass())
            return false;
        EmailTypeDTO other = (EmailTypeDTO) obj;
        if (active != other.active)
            return false;
        if (description == null)
        {
            if (other.description != null)
                return false;
        }
        else if (!description.equals(other.description))
            return false;
        if (id != other.id)
            return false;
        return true;
    }

}
